* querybuilder: move hydrate and persist to QueryTranslator

+++ packages/elcy-core/src/Common/ColumnType.ts
@@ -58,7 +58,7 @@ export type IntervalColumnType = "interval year" // oracle
     | "interval" // postgres
     | "year"; // mysql
 
-export type RowVersionColumnType = "bigint" // all
+export type RowVersionColumnType = "int" // all
     | "rowversion" // mssql
     | "xmin"; // postgres
